refactor(remote-inference): split RemoteInferenceSection into parts
